Keep Kalman PR focused on indicator changes
