+++ CHANGELOG.md
@@ -8,7 +8,6 @@
 - **Staged rule composition for thesis-based exits**: Added `TriggeredRule` so strategies can keep invalidation and timeout rules active while trigger/delegate stages arm follow-on exits such as trailing stops. Stages support finite or unbounded activation windows, delegate priming, explicit reset rules, and position-change resets for reusable target, timeout, and risk-management workflows.
 - **Anchor-aware Elliott tuning harness**: You can now run `ElliottWaveAnchorCalibrationHarness` with the versioned `BTC-anchor-registry-v1.json` registry to replay major BTC tops and bottoms, compare the locked baseline against coarse challenger profiles on top-1/top-3 anchor hits plus `ECE`/`Brier`/`LogLoss`, and get a deterministic promote-or-retain bundle that also sanity-checks the selected profile on ETH/USD and S&P 500 data.
 - **Backtesting throughput comparisons are now reproducible**: `BacktestPerformanceTuningHarness` can write fixed-matrix `matrix_performance.json` artifacts with cells/min, hypotheses/min, hashed host telemetry, and per-cell timings, while `scripts/benchmark-backtest-throughput.sh` compares `HEAD^` and `HEAD` on the same dataset/spec/host for focused tuning work (`CF-170`).
-- **Reusable performance experiment runner**: `PerformanceExperimentRunner` can run named examples-layer performance experiments, capture `performance.json` plus Markdown summaries with hashed host/JVM metadata, and compare two git refs through `scripts/benchmark-performance-experiment.sh`. The first packaged workload covers `KalmanFilterIndicator` access patterns.
 - **Release automation audit artifacts**: Release workflows now emit structured release dossiers, model catalog preflight data, normalized AI decisions, exact artifact manifests, tag-resolution snapshots, full-build logs, and workflow summaries so release failures can be diagnosed from the exact step that produced them.
 - **Report-producing analysis demos can run on demand or on schedule**: The new standalone `analysis-demo` JUnit tag starts with a live Elliott Wave BTC/USD macro report backed by Coinbase daily candles, plus a dedicated GitHub Actions workflow that can be manually run for provider-qualified instruments like `coinbase:BTC-USD` and uploads the resulting JSON, chart, and cached data artifacts.
 - **Elliott macro-cycle replay has a dedicated runner path**: The BTC macro-cycle replay regression now uses the explicit `elliott-macro-cycle-replay` JUnit tag and a manual-only self-hosted GitHub Actions workflow, keeping the multi-hour replay out of hosted PR verification and report-producing analysis-demo runs.
